Split out the replacer engine into its own file.
Also rework the code to handle exceptions right and not be quite so
miserable in how it iterates over files' content lines.
